rename octaviusErrors to errors
