+++ docs/errorhandling.md
@@ -1,7 +1,7 @@
 # Error Handling
 
 1. Create custom error message using package octaviusErrors while returning the error
-    - `errMsg := octaviusErrors.New(<error-code>, <error>)`
+    - `errMsg := Errors.New(<error-code>, <error>)`
     - for <error-code> refer the below map
         - 0: for No Error
         - 1: for Client
